Setup first ViewController and set it as root view controller. SquaresViewController calls its own loadView method to setup its own root view, and adds that root view as a subview of UIWindow.
