+++ Donnie/AppDelegate.swift
@@ -12,138 +12,38 @@ import UIKit
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
-
-
+    
+    
     func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
         // Override point for customization after application launch.
-        var vc = ViewController()
+        var vc = SquaresViewController()
         self.window = UIWindow(frame: UIScreen.mainScreen().bounds)
         window?.rootViewController = vc
         window?.makeKeyAndVisible()
         
-        self.window?.backgroundColor = UIColor(red: 1, green: 1, blue: 1, alpha: 1)
-        
-        var redRect = UIButton(frame: CGRect(x: 8, y: 32, width: 175.5, height: 175.5))
-        redRect.backgroundColor = UIColor.redColor()
-        redRect.addTarget(self, action: "redButtonTouched:", forControlEvents: UIControlEvents.TouchUpInside)
-        
-        var greenRect = UIButton(frame: CGRect(x: 191.5, y: 32, width: 175.5, height: 175.5))
-        greenRect.backgroundColor = UIColor.greenColor()
-        greenRect.addTarget(self, action: "greenButtonTouched:", forControlEvents: UIControlEvents.TouchUpInside)
-        
-        var blueRect = UIButton(frame: CGRect(x: 8, y: 215.5, width: 175.5, height: 175.5))
-        blueRect.backgroundColor = UIColor.blueColor()
-        blueRect.addTarget(self, action: "blueButtonTouched:", forControlEvents: UIControlEvents.TouchUpInside)
-        
-        var yellowRect = UIButton(frame: CGRect(x: 191.5, y: 215.5, width: 175.5, height: 175.5))
-        yellowRect.backgroundColor = UIColor.yellowColor()
-        yellowRect.addTarget(self, action: "yellowButtonTouched:", forControlEvents: UIControlEvents.TouchUpInside)
-        
-        self.window?.addSubview(redRect)
-        self.window?.addSubview(greenRect)
-        self.window?.addSubview(blueRect)
-        self.window?.addSubview(yellowRect)
-        
-        var redLabel = UILabel(frame: CGRect(x: 8, y: 8, width: 0, height: 0))
-        redLabel.text = "Red Square"
-        redLabel.backgroundColor = UIColor.lightGrayColor()
-        redLabel.sizeToFit()
-        redLabel.textAlignment = NSTextAlignment.Left
-        
-        var greenLabel = UILabel(frame: CGRect(x: 8, y: 8, width: 0, height: 0))
-        greenLabel.text = "Green Square"
-        greenLabel.backgroundColor = UIColor.lightGrayColor()
-        greenLabel.sizeToFit()
-        greenLabel.textAlignment = NSTextAlignment.Left
-        
-        var blueLabel = UILabel(frame: CGRect(x: 8, y: 8, width: 0, height: 0))
-        blueLabel.text = "Blue Square"
-        blueLabel.backgroundColor = UIColor.lightGrayColor()
-        blueLabel.sizeToFit()
-        blueLabel.textAlignment = NSTextAlignment.Left
-        
-        var yellowLabel = UILabel(frame: CGRect(x: 8, y: 8, width: 0, height: 0))
-        yellowLabel.text = "Yellow Square"
-        yellowLabel.backgroundColor = UIColor.lightGrayColor()
-        yellowLabel.sizeToFit()
-        yellowLabel.textAlignment = NSTextAlignment.Left
-        
-        var label = UILabel(frame: CGRect(x: 8, y: 397, width: 359, height: 20))
-        label.text = "Hello World"
-        label.backgroundColor = UIColor.lightGrayColor()
-        label.textAlignment = NSTextAlignment.Center
-        
-        redRect.addSubview(redLabel)
-        greenRect.addSubview(greenLabel)
-        blueRect.addSubview(blueLabel)
-        yellowRect.addSubview(yellowLabel)
-        self.window?.addSubview(label)
-        
-        var redImage = UIImageView(frame: CGRect(x: 62.75, y: 62.75, width: 50, height: 50))
-        redImage.image = UIImage(named: "Star.png")
-        redRect.addSubview(redImage)
-        
-        var greenImage = UIImageView(frame: CGRect(x: 62.75, y: 62.75, width: 50, height: 50))
-        greenImage.image = UIImage(named: "Star.png")
-        greenRect.addSubview(greenImage)
-        
-        var blueImage = UIImageView(frame: CGRect(x: 62.75, y: 62.75, width: 50, height: 50))
-        blueImage.image = UIImage(named: "Star.png")
-        blueRect.addSubview(blueImage)
-        
-        var yellowImage = UIImageView(frame: CGRect(x: 62.75, y: 62.75, width: 50, height: 50))
-        yellowImage.image = UIImage(named: "Star.png")
-        yellowRect.addSubview(yellowImage)
-        
-        var myButton = UIButton(frame: CGRect(x: 8, y: 433, width: 359, height: 20))
-        myButton.setTitle("Start", forState: UIControlState.Normal)
-        myButton.setTitleColor(UIColor.blueColor(), forState: UIControlState.Normal)
-        self.window?.addSubview(myButton)
-        myButton.addTarget(self, action: "buttonTouched:", forControlEvents: UIControlEvents.TouchUpInside)
-        
         return true
     }
-
+    
     func applicationWillResignActive(application: UIApplication) {
         // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
         // Use this method to pause ongoing tasks, disable timers, and throttle down OpenGL ES frame rates. Games should use this method to pause the game.
     }
-
+    
     func applicationDidEnterBackground(application: UIApplication) {
         // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
         // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
     }
-
+    
     func applicationWillEnterForeground(application: UIApplication) {
         // Called as part of the transition from the background to the inactive state; here you can undo many of the changes made on entering the background.
     }
-
+    
     func applicationDidBecomeActive(application: UIApplication) {
         // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
     }
-
+    
     func applicationWillTerminate(application: UIApplication) {
         // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
-    }
-    
-    func buttonTouched(button:UIButton) {
-        println("The user touched the start button")
-    }
-    
-    func redButtonTouched(button:UIButton) {
-        println("The user touched this square: \(button)")
-    }
-    
-    func greenButtonTouched(button:UIButton) {
-        println("The user touched this square: \(button)")
-    }
-    
-    func blueButtonTouched(button:UIButton) {
-        println("The user touched this square: \(button)")
-    }
-    
-    func yellowButtonTouched(button:UIButton) {
-        println("The user touched this square: \(button)")
     }
 }
