--- Donnie/ViewController.swift
@@ -1,26 +1,0 @@
-//
-//  ViewController.swift
-//  Donnie
-//
-//  Created by Gilman Tolle on 5/17/15.
-//  Copyright (c) 2015 Donnie Wang. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-        
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
